relocate TiFlashTestEnv
