+++ dbms/src/Debug/MockComputeServerManager.cpp
@@ -15,10 +15,10 @@
 #include <Common/FmtUtils.h>
 #include <Debug/MockComputeServerManager.h>
 #include <Debug/MockStorage.h>
+#include <Debug/TiFlashTestEnv.h>
 #include <Flash/Mpp/MPPTaskManager.h>
 #include <Interpreters/Context.h>
 #include <Storages/KVStore/TMTContext.h>
-#include <TestUtils/TiFlashTestEnv.h>
 
 #include <chrono>
 #include <thread>
