+++ dbms/src/Interpreters/tests/gtest_operator_spill_context.cpp
@@ -14,6 +14,7 @@
 
 #include <Common/Logger.h>
 #include <Core/FineGrainedOperatorSpillContext.h>
+#include <Debug/TiFlashTestEnv.h>
 #include <IO/Encryption/MockKeyManager.h>
 #include <IO/FileProvider/FileProvider.h>
 #include <Interpreters/AggSpillContext.h>
@@ -21,7 +22,6 @@
 #include <Interpreters/SortSpillContext.h>
 #include <Poco/File.h>
 #include <TestUtils/TiFlashTestBasic.h>
-#include <TestUtils/TiFlashTestEnv.h>
 #include <gtest/gtest.h>
 
 namespace DB
