+++ dbms/src/Server/CMakeLists.txt
@@ -28,12 +28,9 @@ option(ENABLE_TIFLASH_CHECKPOINTTOOL "Enable checkpoint tool: tools to debug che
 
 configure_file (config_tools.h.in ${CMAKE_CURRENT_BINARY_DIR}/config_tools.h)
 
-add_library(server_for_test
-    FlashGrpcServerHolder.cpp
-)
-
-target_link_libraries (server_for_test PUBLIC tiflash_common_io tiflash_storages_system)
-target_include_directories (server_for_test PRIVATE ${TiFlash_SOURCE_DIR}/contrib/grpc/)
+add_library(grpc_server_holder FlashGrpcServerHolder.cpp)
+target_link_libraries (grpc_server_holder PUBLIC tiflash_common_io tiflash_storages_system)
+target_include_directories (grpc_server_holder PRIVATE ${TiFlash_SOURCE_DIR}/contrib/grpc/)
 
 add_library (tiflash-server-lib
     CertificateReloader.cpp
@@ -52,8 +49,8 @@ add_library (tiflash-server-lib
 
 target_link_libraries(tiflash-server-lib PRIVATE ${TIFLASH_PROXY_LIBRARY})
 
-target_link_libraries (tiflash-server-lib PUBLIC server_for_test tiflash_common_io daemon tiflash_storages_system tiflash_functions tiflash_aggregate_functions tiflash_table_functions)
-target_include_directories (tiflash-server-lib PUBLIC ${TiFlash_SOURCE_DIR}/libs/libdaemon/include)
+target_link_libraries (tiflash-server-lib PRIVATE grpc_server_holder tiflash_common_io daemon tiflash_storages_system tiflash_functions tiflash_aggregate_functions tiflash_table_functions)
+target_include_directories (tiflash-server-lib PRIVATE ${TiFlash_SOURCE_DIR}/libs/libdaemon/include)
 target_include_directories (tiflash-server-lib PRIVATE ${TiFlash_SOURCE_DIR}/contrib/GmSSL/include)
 
 add_library (tiflash-client-lib Client.cpp)
