+++ dbms/src/Server/Server.h
@@ -14,7 +14,6 @@
 
 #pragma once
 
-#include <Server/FlashGrpcServerHolder.h>
 #include <Server/IServer.h>
 #include <Server/ServerInfo.h>
 #include <daemon/BaseDaemon.h>
