+++ dbms/src/Storages/Page/V3/tests/gtest_versioned_entries.cpp
@@ -25,13 +25,9 @@
 #include <TestUtils/MockDiskDelegator.h>
 #include <TestUtils/TiFlashStorageTestBasic.h>
 #include <TestUtils/TiFlashTestBasic.h>
-#include <TestUtils/TiFlashTestEnv.h>
 #include <common/logger_useful.h>
 #include <common/types.h>
 #include <fmt/format.h>
-
-#include <iterator>
-#include <memory>
 
 #pragma GCC diagnostic push
 #ifdef __clang__
