+++ dbms/src/Storages/Page/tools/PageCtl/PageStorageCtlV3.cpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include <Common/Checksum.h>
+#include <Debug/TiFlashTestEnv.h>
 #include <IO/Encryption/MockKeyManager.h>
 #include <IO/FileProvider/FileProvider.h>
 #include <Interpreters/Context.h>
@@ -31,7 +32,6 @@
 #include <Storages/Page/WriteBatchImpl.h>
 #include <Storages/PathPool.h>
 #include <TestUtils/MockDiskDelegator.h>
-#include <TestUtils/TiFlashTestEnv.h>
 #include <boost_wrapper/program_options.h>
 #include <common/types.h>
