+++ dbms/src/Storages/S3/tests/gtest_s3client.cpp
@@ -14,9 +14,9 @@
 
 #include <Common/FailPoint.h>
 #include <Common/Logger.h>
+#include <Debug/TiFlashTestEnv.h>
 #include <Storages/S3/S3Common.h>
 #include <TestUtils/TiFlashTestBasic.h>
-#include <TestUtils/TiFlashTestEnv.h>
 #include <aws/s3/model/ListObjectsV2Request.h>
 #include <gtest/gtest.h>
