+++ dbms/src/TestUtils/CMakeLists.txt
@@ -18,7 +18,7 @@ add_headers_and_sources(test_util .)
 list(REMOVE_ITEM test_util_sources "bench_dbms_main.cpp" "gtests_dbms_main.cpp")
 
 add_library(test_util_gtest_main ${test_util_headers} ${test_util_sources} gtests_dbms_main.cpp)
-target_link_libraries(test_util_gtest_main dbms gtest_main tiflash_aggregate_functions server_for_test)
+target_link_libraries(test_util_gtest_main dbms gtest_main tiflash_aggregate_functions grpc_server_holder)
 target_compile_definitions(test_util_gtest_main PRIVATE DBMS_PUBLIC_GTEST)
 
 add_library(test_util_bench_main ${test_util_headers} ${test_util_sources} bench_dbms_main.cpp)
