+++ dbms/src/TestUtils/MPPTaskTestUtils.cpp
@@ -14,7 +14,6 @@
 #include <Debug/MockComputeServerManager.h>
 #include <Debug/MockStorage.h>
 #include <Debug/dbgQueryExecutor.h>
-#include <Server/FlashGrpcServerHolder.h>
 #include <Server/MockComputeClient.h>
 #include <Storages/KVStore/TMTContext.h>
 #include <Storages/KVStore/Types.h>
